+++ src/vaxflux/__init__.py
@@ -2,33 +2,28 @@
 
 __all__ = (
     "Covariate",
+    "CovariateCategories",
     "Curve",
+    "DateRange",
     "Implementation",
     "Intervention",
     "LogisticCurve",
     "PartiallyPooledGaussianCovariate",
     "PooledGaussianCovariate",
+    "SeasonRange",
     "SeasonVaryingPartiallyPooledGaussianCovariate",
     "VaxfluxInferenceData",
     "VaxfluxModel",
-    "covariates",
-    "curves",
+    "daily_date_ranges",
     "data",
-    "dates",
-    "interventions",
-    "uptake",
 )
 __version__ = "0.2.0"
 
 
 from vaxflux import (
-    covariates,
-    curves,
     data,
-    dates,
-    interventions,
-    uptake,
 )
+from vaxflux._covariate_categories import CovariateCategories
 from vaxflux._covariates import (
     Covariate,
     PartiallyPooledGaussianCovariate,
@@ -36,6 +31,7 @@ from vaxflux._covariates import (
     SeasonVaryingPartiallyPooledGaussianCovariate,
 )
 from vaxflux._curves import Curve, LogisticCurve
+from vaxflux._dates import DateRange, SeasonRange, daily_date_ranges
 from vaxflux._interventions import Implementation, Intervention
 from vaxflux._vaxflux_inference_data import VaxfluxInferenceData
 from vaxflux._vaxflux_model import VaxfluxModel